+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Unreleased
+
+### General
+-
+
+### Client
+- Feat: 添付された3DモデルとGaussian Splatをクリックしてプレビューできるように
+
+### Server
+-
+
 ## 2026.7.0
 
 ### Note